% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampbias.R
\docType{data}
\name{area_example}
\alias{area_example}
\title{Example Dataset for a Custom Study Area}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 2 rows and 1 columns.
}
\usage{
area_example
}
\description{
An example of the format needed to provide custom areas for \code{\link{calculate_bias}}
}
\examples{

data(area_example)

}
\keyword{datasets}
