% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampbias.R
\docType{data}
\name{borneo}
\alias{borneo}
\title{Borneo}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 3 columns.
}
\usage{
borneo
}
\description{
The outline of Borneo, as example data for the user-defined study area option
of \code{\link{calculate_bias}}. From \url{https://www.naturalearthdata.com}.
}
\examples{

data(borneo)

}
\keyword{datasets}
