% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampbias.R
\docType{data}
\name{ea_raster}
\alias{ea_raster}
\title{Equal Area Raster}
\format{
An object of class \code{PackedSpatRaster} of length 1.
}
\usage{
ea_raster
}
\description{
An example for an global equal area raster (in Behrmann projection) for the
format needed for a custom grid provided to \code{\link{calculate_bias}}.
}
\examples{

data(ea_raster)
ea_raster <- terra::unwrap(ea_raster)
}
\keyword{datasets}
