\name{exampledata_large}
\alias{exampledata_large}
\docType{data}
\title{
Plant Species Distributions in West Africa (large)
}
\description{
An example dataset with 15,000 plant records from West Africa
}
\usage{data("exampledata_large")}
\format{
  A data frame with 15000 observations on the following 3 variables.
  \describe{
    \item{\code{species}}{a factor with species names}
    \item{\code{longitude}}{a numeric vector}
    \item{\code{latitude}}{a numeric vector}
  }
}

\source{
Schmidt, Zizka et al 2015
}

\examples{
data(exampledata)
str(exampledata)
}
\keyword{datasets}
