% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampbias.R
\docType{data}
\name{landmass}
\alias{landmass}
\title{Global Coastlines}
\source{
\url{http://www.naturalearthdata.com/downloads/10m-physical-vectors/}
}
\description{
A \code{SpatialPolygonsDataFrame} with global coastlines.
}
\examples{

data("landmass")


}
\keyword{datasets}
