\name{roads}
\alias{roads}
\docType{data}
\title{
Global Network of Major Roads
}
\description{
A gazetteer of major global roads as \code{SpatialPolygonsDataFrame}.
}
\usage{data("roads")}

\source{
\url{http://www.naturalearthdata.com/downloads/10m-cultural-vectors/}
}

\examples{
data(roads)
}
\keyword{datasets}
