\name{waterbodies}
\alias{waterbodies}
\docType{data}
\title{
Global Major Waterbodies
}
\description{
A \code{SpatialPolygonsDataFrame} with the centerlines of major global waterbodies.
}
\usage{data("waterbodies")}

\source{
\url{http://www.naturalearthdata.com/downloads/10m-physical-vectors/}
}

\examples{
data(waterbodies)
plot(waterbodies)
}
\keyword{datasets}
