% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampbias.R
\docType{data}
\name{airports}
\alias{airports}
\title{Global Airport Coordinates}
\format{
A \code{sf} object with 891 observations on the following 4 variables.
\describe{ \item{code}{a factor, IATA-3 code.} \item{name}{a
character vector, airport names.} \item{geometry}{sfc_POINT object.} }
}
\source{
\url{http://www.naturalearthdata.com/}
}
\description{
A gazetteer of major global airport locations.
}
\examples{

data(airports)
str(airports)

}
\keyword{datasets}
