% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampbias.R
\docType{data}
\name{borneo}
\alias{borneo}
\title{Borneo}
\format{
An object of class \code{SpatialPolygonsDataFrame} with 1 rows and 2 columns.
}
\usage{
borneo
}
\description{
The outline of Borneo, as example data for the user-defined study area option of \code{\link{calculate_bias}}.
From \url{www.naturalearth.org}.
}
\examples{

data(borneo)

}
\keyword{datasets}
