% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampbias.R
\docType{data}
\name{cities}
\alias{cities}
\title{Global Major Populated Places}
\format{A data frame with 7322 observations on the following 4 variables.
\describe{ \item{list("nameascii")}{a factor; city name.}
\item{list("adm0_a3")}{a factor; ISO-3 code of country name.}
\item{list("longitude")}{a numeric vector.} \item{list("latitude")}{a
numeric vector.} }}
\source{
\url{http://www.naturalearthdata.com/downloads/10m-cultural-vectors/}
}
\description{
A gazetteer of global country and province capitals, as well as major cities
and smaller cities in less populated areas.
}
\examples{

data(cities)
str(cities)

}
\keyword{datasets}
