% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampbias.R
\docType{data}
\name{ecoregion_example}
\alias{ecoregion_example}
\title{Detailed Example for a Custom Study Area}
\format{
An object of class \code{SpatialPolygonsDataFrame} with 7 rows and 21 columns.
}
\source{
\url{https://www.worldwildlife.org/publications/terrestrial-ecoregions-of-the-world}
}
\usage{
ecoregion_example
}
\description{
An example of the format needed to provide custom areas for \code{\link{calculate_bias}}
based on a publicly available set of global ecoregions.
}
\examples{

data(ecoregion_example)

}
\keyword{datasets}
