\name{landmass}
\alias{landmass}
\docType{data}
\title{
Global Coastlines
}
\description{
A \code{SpatialPolygonsDataFrame} with global coastlines.
}
\usage{data("landmass")}

\source{
\url{http://www.naturalearthdata.com/downloads/10m-physical-vectors/}
}

\examples{
data("landmass")
plot(landmass)
}
\keyword{datasets}
