% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampbias.R
\docType{data}
\name{roads}
\alias{roads}
\title{Global Network of Major Roads}
\source{
\url{http://www.naturalearthdata.com/downloads/10m-cultural-vectors/}
}
\description{
A gazetteer of major global roads as \code{SpatialPolygonsDataFrame}.
}
\examples{

data(roads)

}
\keyword{datasets}
