% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampbias.R
\docType{data}
\name{waterbodies}
\alias{waterbodies}
\title{Global Major Waterbodies}
\source{
\url{http://www.naturalearthdata.com/downloads/10m-physical-vectors/}
}
\description{
A \code{SpatialPolygonsDataFrame} with the centerlines of major global
waterbodies.
}
\examples{

data(waterbodies)

}
\keyword{datasets}
